// app/interfaces/MainComponent/AccessSection.ts
export interface IAccessSection {
    _id: string;
    id?: string;
    pagetitle?: string;
    title?: string;
    image?: string;
    description?: string;
}

export interface IAccessSectionInput {
    id?: string;
    pagetitle?: string;
    title?: string;
    image?: string;
    description?: string;
}